+++ src/__tests__/core/workspaceFolder.test.ts
@@ -30,8 +30,7 @@ beforeAll(async () => {
   let userConfigFile = path.join(process.env.COC_VIMCONFIG, 'coc-settings.json')
   configurations = new Configurations(userConfigFile, {
     $removeConfigurationOption: () => {},
-    $updateConfigurationOption: () => {},
-    workspaceConfigFile: ''
+    $updateConfigurationOption: () => {}
   })
   workspaceFolder = new WorkspaceFolderController(configurations)
 })