+++ history.md
@@ -1,3 +1,8 @@
+# 2022-02-17
+
+- Change workspace configuration throw error when workspace folder can't be
+  resolved.
+
 # 2022-02-15
 
 - Add `events.race`.